allow rendering full wiki syntax
This adds the option to parse full wiki syntax from the queried result.
The column needs to be selected with an alias ending in _wiki.

Headers will be output as paragraphs with bold formatting, not as
headers!

For performance reasons the parsing will not initialize a new rendering
context but inject the results into the current renderer. This has the
additional advantage of being compatible with any renderer (like PDF or
ODT). It might still break on certain plugins.