placeholders to access local sqlit dbs
When using the plugin to access existing SQLite databases (eg from
struct) it is helpful to be able to reference the local data directory.

This becomes doubly helpful when using the plugin in a farmer setup and
wanting to define aliases in the farmer to be inherited by animals. The
placeholders will reference the animal's data dir automatically.